instead of performing database lookups. This gives a 14% speed-up when syncing the toilet wallet:

| operation | before | after | after / before |
| --- | --- | --- | --- |
| wallet sync | 83.7s | 71.96s | 85.97% |